Add bibliography
We've already got the one citation about the BIOS memory map in the
linker script. This is also copied from zero, but most of the resources
have been pulled out for the time being. We don't have any citations
from the Intel manual *yet*, but that will probably happen, so we leave
them there.