Fixes to the linker script
Rust seems to put each module in it's own section. The impact of this is
that doing e.g. *(.bss) doesn't catch bss stuff in rust code, just the
stuff we've declared in asm and the linker script itself. so this adds
*(.bss.*) and friends. Previously this was causing a problem where
kend == &BOOT_IDT, which was bad.

This also puts the align statements at the *start* of the sections,
rather than the end.

Note that this is still not giving me the memory info -- but at least it
doesn't change when I don't call init_idt(), so that's something.